Fix length check in TCP and UDP into_cstruct functions
They were checking that the buffer was big enough for the headers and
the payload, but they don't write the payload and there's no reason to
require it to be written to the same buffer.

Also, change the IPv4 marshal functions to take the payload length
rather than the payload itself, so that the caller can provide the
payload in multiple buffers. This also makes it clear that the function
does not actually write the payload.